% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_init_temp}
\alias{auto_init_temp}
\title{Automatically determine initial temperature for simulated annealing}
\arguments{
\item{sol_mat}{Current solution matrix}

\item{snp_pairs}{SNP pair data (male_idx, female_idx, target_cor)}

\item{female_swap_idx}{Columns to swap in solution}

\item{num_samples}{Number of random moves to sample}

\item{accept_ratio}{Target initial acceptance ratio (default: 0.8)}
}
\value{
Recommended initial temperature
}
\description{
Automatically determine initial temperature for simulated annealing
}
\keyword{internal}
