% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_delta}
\alias{compute_delta}
\title{Compute update vector for correlation vector of assortative trait pairs.}
\usage{
compute_delta(sol, swap, phi_m, phi_f)
}
\arguments{
\item{sol}{The current state matrix}

\item{swap}{Indices of swapped pairs}

\item{phi_m}{Male phenotype indices}

\item{phi_f}{Female phenotype indices}
}
\value{
Update vector for correlation vector of assortative trait pairs
}
\description{
Compute update vector for correlation vector of assortative trait pairs.
}
