% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{compute_sib_cor}
\alias{compute_sib_cor}
\title{Calculate the genetic correlation between siblings in a population}
\usage{
compute_sib_cor(config, population, phenotype_name)
}
\arguments{
\item{config}{Configuration list with simulation parameters}

\item{population}{Data frame containing the population with siblings in
consecutive pairs}

\item{phenotype_name}{Name of the phenotype column (optional)}
}
\value{
A list containing the genetic correlation and additional statistics
}
\description{
Calculate the genetic correlation between siblings in a population
}
