% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__environmental_cor}
\alias{.SimulationConfig__environmental_cor}
\title{Get the phenotype environmental component correlation matrix.}
\usage{
.SimulationConfig__environmental_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype environmental component correlation matrix.
}
