% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__genetic_cor}
\alias{.SimulationConfig__genetic_cor}
\title{Get the phenotype genetic component correlation matrix.}
\usage{
.SimulationConfig__genetic_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype genetic component correlation matrix.
}
