% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__h2_environmental}
\alias{.SimulationConfig__h2_environmental}
\title{Get the phenotype environmental component variances.}
\usage{
.SimulationConfig__h2_environmental(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype environmental component variances.
}
