% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__h2_genetic}
\alias{.SimulationConfig__h2_genetic}
\title{Get the narrow-sense phenotype heritabilities.}
\usage{
.SimulationConfig__h2_genetic(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the narrow-sense phenotype heritabilities.
}
