% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__h2_vertical}
\alias{.SimulationConfig__h2_vertical}
\title{Get the phenotype vertical component variances.}
\usage{
.SimulationConfig__h2_vertical(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype vertical component variances.
}
