% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__locus_maf}
\alias{.SimulationConfig__locus_maf}
\title{Get the locus MAF vector.}
\usage{
.SimulationConfig__locus_maf(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the locus MAF vector.
}
