% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__locus_mutation}
\alias{.SimulationConfig__locus_mutation}
\title{Get the locus mutation probabilities.}
\usage{
.SimulationConfig__locus_mutation(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the locus mutation probabilities.
}
