% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__locus_recombination}
\alias{.SimulationConfig__locus_recombination}
\title{Get the locus recombination probabilities.}
\usage{
.SimulationConfig__locus_recombination(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the locus recombination probabilities.
}
