% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__mate_cor}
\alias{.SimulationConfig__mate_cor}
\title{Get the mating correlation matrix.}
\usage{
.SimulationConfig__mate_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the mating correlation matrix.
}
