% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__metrics}
\alias{.SimulationConfig__metrics}
\title{Configure simulation metrics}
\usage{
.SimulationConfig__metrics(ptr, metrics)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{metrics}{List of metric objects to attach to the simulation.}
}
\value{
Self, for method chaining.
}
\description{
Configure simulation metrics
}
