% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__n_causal_loci}
\alias{.SimulationConfig__n_causal_loci}
\title{Get the phenotype names.}
\usage{
.SimulationConfig__n_causal_loci(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype names.
}
