% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__n_generations}
\alias{.SimulationConfig__n_generations}
\title{Get the number of generations.}
\usage{
.SimulationConfig__n_generations(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the number of generations.
}
