% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__n_individuals}
\alias{.SimulationConfig__n_individuals}
\title{Get the number of individuals.}
\usage{
.SimulationConfig__n_individuals(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the number of individuals.
}
