% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__n_iterations}
\alias{.SimulationConfig__n_iterations}
\title{Get the maximum number of annealing iterations.}
\usage{
.SimulationConfig__n_iterations(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the maximum number of annealing iterations.
}
