% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__n_phenotypes}
\alias{.SimulationConfig__n_phenotypes}
\title{Get the number of phenotypes.}
\usage{
.SimulationConfig__n_phenotypes(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the number of phenotypes.
}
