% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__new}
\alias{.SimulationConfig__new}
\title{Create a new SimulationConfig object reference}
\usage{
.SimulationConfig__new()
}
\value{
A new SimulationConfig instance
}
\description{
Create a new SimulationConfig object reference
}
