% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__output_dir}
\alias{.SimulationConfig__output_dir}
\title{Get the output directory.}
\usage{
.SimulationConfig__output_dir(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the output directory.
}
