% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__phenome}
\alias{.SimulationConfig__phenome}
\title{Configure phenome parameters.}
\usage{
.SimulationConfig__phenome(
  ptr,
  n_phenotypes,
  names,
  n_causal_loci,
  h2_genetic,
  h2_environmental,
  h2_vertical,
  genetic_cor,
  environmental_cor
)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{n_phenotypes}{Number of phenotypes modelled in the population.}

\item{names}{Vector of phenotype names.}

\item{n_causal_loci}{Number of causal loci for each phenotype.}

\item{h2_genetic}{Narrow-sense phenotype heritabilities.}

\item{h2_environmental}{Phenotype environmental component variances.}

\item{h2_vertical}{Phenotype vertical component variances.}
}
\value{
Self, for method chaining
}
\description{
Configure phenome parameters.
}
