% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__phenotype_names}
\alias{.SimulationConfig__phenotype_names}
\title{Get the phenotype names.}
\usage{
.SimulationConfig__phenotype_names(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the phenotype names.
}
