% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__random_mating}
\alias{.SimulationConfig__random_mating}
\title{Specify a random mating regime.}
\usage{
.SimulationConfig__random_mating(ptr)
}
\value{
Self, for method chaining.
}
\description{
Specify a random mating regime.
}
