% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__random_seed}
\alias{.SimulationConfig__random_seed}
\title{Get the RNG seed.}
\usage{
.SimulationConfig__random_seed(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the RNG seed.
}
