% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__run_simulations}
\alias{.SimulationConfig__run_simulations}
\title{Run (multithreaded) simulations}
\usage{
.SimulationConfig__run_simulations(
  ptr,
  n_replicates,
  n_threads,
  summarise = TRUE,
  log_file = FALSE,
  log_level = "info"
)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{n_replicates}{Integer. Number of independent simulation replicates to
run.}

\item{n_threads}{Integer. Number of parallel threads to use for execution.}

\item{summarise}{Logical. Whether to summarize results across replicates
(default: TRUE).}

\item{log_file}{Logical. Whether to write log output to a file (default:
FALSE).}

\item{log_level}{String. Logging verbosity level.}
}
\description{
Run (multithreaded) simulations
}
