% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__temp_decay}
\alias{.SimulationConfig__temp_decay}
\title{Get the annealing temperature decay parameter.}
\usage{
.SimulationConfig__temp_decay(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the annealing temperature decay parameter.
}
