% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__temp_init}
\alias{.SimulationConfig__temp_init}
\title{Get the initial annealing temperature.}
\usage{
.SimulationConfig__temp_init(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the initial annealing temperature.
}
