% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig__tol_inf}
\alias{.SimulationConfig__tol_inf}
\title{Get the annealing tolerance threshold.}
\usage{
.SimulationConfig__tol_inf(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\description{
Get the annealing tolerance threshold.
}
