% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_assortative_mating}
\alias{.SimulationConfig_assortative_mating}
\title{Specify an assortative mating regime}
\usage{
.SimulationConfig_assortative_mating(
  ptr,
  mate_cor,
  tol_inf = NULL,
  n_iterations = NULL,
  temp_init = NULL,
  temp_decay = NULL
)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{mate_cor}{Column-major double trait cross-correlation matrix.}

\item{tol_inf}{Termination threshold for annealing routine.}

\item{n_iterations}{Maximum number of steps performed in annealing routine.}

\item{temp_init}{Initial temperature of annealing routine.}

\item{temp_decay}{Temperature decay parameter of annealing routine.}
}
\value{
Self, for method chaining
}
\description{
Specify an assortative mating regime
}
