% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_environmental_cor}
\alias{.SimulationConfig_environmental_cor}
\title{Get the phenotype environmental component correlation matrix}
\usage{
.SimulationConfig_environmental_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector representing the environmental correlation matrix in
column-major order.
}
\description{
Get the phenotype environmental component correlation matrix
}
