% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_genetic_cor}
\alias{.SimulationConfig_genetic_cor}
\title{Get the phenotype genetic component correlation matrix}
\usage{
.SimulationConfig_genetic_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector representing the genetic correlation matrix in
column-major order.
}
\description{
Get the phenotype genetic component correlation matrix
}
