% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_genome}
\alias{.SimulationConfig_genome}
\title{Configure genome parameters}
\usage{
.SimulationConfig_genome(
  ptr,
  n_loci,
  locus_maf,
  locus_recombination,
  locus_mutation
)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{n_loci}{Number of genetic loci simulated in the population.}

\item{locus_maf}{Vector of locus MAFs.}

\item{locus_recombination}{Vector of locus recombination probabilities.}

\item{locus_mutation}{Vector of locus mutation probabilities.}
}
\value{
Self, for method chaining.
}
\description{
Configure genome parameters
}
