% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_h2_environmental}
\alias{.SimulationConfig_h2_environmental}
\title{Get the phenotype environmental component variances}
\usage{
.SimulationConfig_h2_environmental(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of environmental variance components for each
phenotype.
}
\description{
Get the phenotype environmental component variances
}
