% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_h2_genetic}
\alias{.SimulationConfig_h2_genetic}
\title{Get the narrow-sense phenotype heritabilities}
\usage{
.SimulationConfig_h2_genetic(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of narrow-sense heritabilities for each phenotype.
}
\description{
Get the narrow-sense phenotype heritabilities
}
