% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_h2_vertical}
\alias{.SimulationConfig_h2_vertical}
\title{Get the phenotype vertical component variances}
\usage{
.SimulationConfig_h2_vertical(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of vertical transmission components for each
phenotype.
}
\description{
Get the phenotype vertical component variances
}
