% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_locus_maf}
\alias{.SimulationConfig_locus_maf}
\title{Get the locus MAF vector}
\usage{
.SimulationConfig_locus_maf(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of minor allele frequencies for each locus.
}
\description{
Get the locus MAF vector
}
