% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_locus_mutation}
\alias{.SimulationConfig_locus_mutation}
\title{Get the locus mutation probabilities}
\usage{
.SimulationConfig_locus_mutation(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of mutation probabilities for each locus.
}
\description{
Get the locus mutation probabilities
}
