% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_locus_recombination}
\alias{.SimulationConfig_locus_recombination}
\title{Get the locus recombination probabilities}
\usage{
.SimulationConfig_locus_recombination(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector of recombination probabilities for each locus.
}
\description{
Get the locus recombination probabilities
}
