% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_mate_cor}
\alias{.SimulationConfig_mate_cor}
\title{Get the mating correlation matrix}
\usage{
.SimulationConfig_mate_cor(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric vector representing the mating correlation matrix in
column-major order.
}
\description{
Get the mating correlation matrix
}
