% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_causal_loci}
\alias{.SimulationConfig_n_causal_loci}
\title{Get the number of causal loci for each phenotype}
\usage{
.SimulationConfig_n_causal_loci(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer vector specifying the number of causal loci per phenotype.
}
\description{
Get the number of causal loci for each phenotype
}
