% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_generations}
\alias{.SimulationConfig_n_generations}
\title{Get the number of generations}
\usage{
.SimulationConfig_n_generations(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the number of generations to simulate.
}
\description{
Get the number of generations
}
