% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_individuals}
\alias{.SimulationConfig_n_individuals}
\title{Get the number of individuals}
\usage{
.SimulationConfig_n_individuals(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the population size.
}
\description{
Get the number of individuals
}
