% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_iterations}
\alias{.SimulationConfig_n_iterations}
\title{Get the maximum number of annealing iterations}
\usage{
.SimulationConfig_n_iterations(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the maximum number of annealing iterations.
}
\description{
Get the maximum number of annealing iterations
}
