% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_loci}
\alias{.SimulationConfig_n_loci}
\title{Get the number of loci}
\usage{
.SimulationConfig_n_loci(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the number of genetic loci.
}
\description{
Get the number of loci
}
