% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_n_phenotypes}
\alias{.SimulationConfig_n_phenotypes}
\title{Get the number of phenotypes}
\usage{
.SimulationConfig_n_phenotypes(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the number of phenotypic traits.
}
\description{
Get the number of phenotypes
}
