% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_new}
\alias{.SimulationConfig_new}
\title{Create a new SimulationConfig object reference}
\usage{
.SimulationConfig_new()
}
\value{
A new SimulationConfig instance.
}
\description{
Create a new SimulationConfig object reference
}
