% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_output_dir}
\alias{.SimulationConfig_output_dir}
\title{Get the output directory}
\usage{
.SimulationConfig_output_dir(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
String specifying the output directory path.
}
\description{
Get the output directory
}
