% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_phenotype_names}
\alias{.SimulationConfig_phenotype_names}
\title{Get the phenotype names}
\usage{
.SimulationConfig_phenotype_names(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Character vector of phenotype names.
}
\description{
Get the phenotype names
}
