% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_random_seed}
\alias{.SimulationConfig_random_seed}
\title{Get the RNG seed}
\usage{
.SimulationConfig_random_seed(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Integer specifying the random number generator seed.
}
\description{
Get the RNG seed
}
