% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_simulation}
\alias{.SimulationConfig_simulation}
\title{Configure simulation parameters}
\usage{
.SimulationConfig_simulation(
  ptr,
  n_generations,
  n_individuals,
  output_dir,
  random_seed
)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}

\item{n_generations}{Integer. Number of generations to simulate.}

\item{n_individuals}{Integer. Number of individuals in the population.}

\item{output_dir}{String. Output directory to write data out to.}

\item{random_seed}{Integer. Seed used by random number generator.}
}
\value{
Self, for method chaining.
}
\description{
Configure simulation parameters
}
