% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_temp_decay}
\alias{.SimulationConfig_temp_decay}
\title{Get the annealing temperature decay parameter}
\usage{
.SimulationConfig_temp_decay(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric value specifying the temperature decay rate.
}
\description{
Get the annealing temperature decay parameter
}
