% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_temp_init}
\alias{.SimulationConfig_temp_init}
\title{Get the initial annealing temperature}
\usage{
.SimulationConfig_temp_init(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric value specifying the initial annealing temperature.
}
\description{
Get the initial annealing temperature
}
