% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationConfig_tol_inf}
\alias{.SimulationConfig_tol_inf}
\title{Get the annealing tolerance threshold}
\usage{
.SimulationConfig_tol_inf(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationConfig instance.}
}
\value{
Numeric value specifying the termination threshold for the
annealing routine.
}
\description{
Get the annealing tolerance threshold
}
