% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationResults_load}
\alias{.SimulationResults_load}
\title{Load a summarised metric into the binding layer.}
\usage{
.SimulationResults_load(ptr, metric_name)
}
\arguments{
\item{ptr}{An external pointer to a SimulationResults instance.}

\item{metric_name}{Name of the metric to be loaded into the environment.}
}
\value{
A DataFrame comprising the summarised results data for the specified
metric.
}
\description{
Load a summarised metric into the binding layer.
}
