% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationResults_metric_names}
\alias{.SimulationResults_metric_names}
\title{Get metric names}
\usage{
.SimulationResults_metric_names(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationResults instance.}
}
\value{
A vector of the metric names recognised by the results instance.
}
\description{
Get metric names
}
