% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationResults_new}
\alias{.SimulationResults_new}
\title{Create a new SimulationResults object reference}
\usage{
.SimulationResults_new(output_dir)
}
\value{
A new SimulationResults instance.
}
\description{
Create a new SimulationResults object reference
}
