% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationResults_save}
\alias{.SimulationResults_save}
\title{Save summarised results to disk}
\usage{
.SimulationResults_save(ptr, metrics = NULL, out_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{ptr}{An external pointer to a SimulationResults instance.}

\item{metrics}{A list of metric names to be saved to file.}

\item{out_dir}{Output directory to write data out to.}

\item{overwrite}{Specifies whether directory should be overwritten.}
}
\description{
Save summarised results to disk
}
