% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.SimulationResults_summarise}
\alias{.SimulationResults_summarise}
\title{Perform summarisation on replicate metric data}
\usage{
.SimulationResults_summarise(ptr)
}
\arguments{
\item{ptr}{An external pointer to a SimulationResults instance.}
}
\description{
Perform summarisation on replicate metric data
}
