% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dpsi}
\alias{dpsi}
\title{Compute the energy differential between proposal state and current state}
\usage{
dpsi(rho_sol, delta, rho)
}
\value{
Energy differential between proposed state and current state
}
\description{
Compute the energy differential between proposal state and current state
}
