% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{from_template}
\alias{from_template}
\title{Create a configuration template}
\usage{
from_template(template_path)
}
\arguments{
\item{template_path}{Path where configuration template is created}
}
\description{
Create a configuration template
}
