% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generate_matching}
\alias{generate_matching}
\title{Generate an approximately optimal mate matching}
\usage{
generate_matching(config, population, collect_metrics = FALSE)
}
\arguments{
\item{config}{Configuration file with simulation parameters}

\item{population}{Population data frame}

\item{collect_metrics}{Whether to collect metrics from simulated annealing
routine to evaluate performance}
}
\value{
List with init_sol matrix, psi_vec, and female_swap_idx
}
\description{
Generate an approximately optimal mate matching
}
