% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generate_phenotype}
\alias{generate_phenotype}
\title{Generate phenotype values from genotypes and environmental effects}
\usage{
generate_phenotype(config, snp_matrix)
}
\arguments{
\item{config}{Configuration list with simulation parameters}

\item{snp_matrix}{Matrix of SNP genotypes (0, 1, or 2 copies of alt allele)}
}
\value{
A matrix with one column containing the phenotype values
}
\description{
Generate phenotype values from genotypes and environmental effects
}
