% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{init_env_variance}
\alias{init_env_variance}
\title{Calibrate the environmental variance so the initial heritabilit
of the phenotype matches the specified value in the config}
\usage{
init_env_variance(config, snp_matrix)
}
\description{
Calibrate the environmental variance so the initial heritabilit
of the phenotype matches the specified value in the config
}
