% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_phenotype}
\alias{is_phenotype}
\title{Validate a phenotype entry}
\usage{
is_phenotype(config, x)
}
\arguments{
\item{config}{Parsed configuration object}

\item{x}{A list representing a phenotype with required fields}
}
\description{
Ensures required keys and valid SNPs/effect values.
}
