% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load a simulation configuration file}
\usage{
load_config(config_path)
}
\arguments{
\item{config_path}{Path to configuration file}
}
\value{
A list object with data processed from the configuration file
}
\description{
Validates and processes data from configuration file to be used
in the simulation.
}
