% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{Metric}
\alias{Metric}
\title{Metric Specification}
\description{
An R6 class for specifying metrics to collect during genetic simulations.
Metrics are statistics calculated and recorded at each generation, such as
heritabilities, phenotypic component means and variances, correlations, and
mate cross-correlations.
}
\details{
Metric objects are created using convenience functions:
\itemize{
\item \code{pheno_h2()} - Phenotype heritabilities
\item \code{pheno_comp_mean()} - Phenotype component means
\item \code{pheno_comp_var()} - Phenotype component variances
\item \code{pheno_comp_cor()} - Phenotype component correlations
\item \code{pheno_comp_xcor()} - Between-mate phenotype component cross-correlations
\item \code{pheno_latent_h2()} - Latent phenotype heritabilities
\item \code{pheno_latent_comp_mean()} - Latent phenotype component means
\item \code{pheno_latent_comp_var()} - Latent phenotype component variances
\item \code{pheno_latent_comp_cor()} - Latent phenotype component correlations
\item \code{pheno_latent_comp_xcor()} - Between-mate latent phenotype cross-correlations
\item \code{custom_metric()} - User-defined metrics
}
}
\examples{
\dontrun{
sim <- Simulation$new()
sim$simulation(
  n_generations = 100,
  n_individuals = 1000,
  output_dir = "output",
  random_seed = 42
)$metrics(
  pheno_h2(),
  pheno_comp_mean("genetic"),
  pheno_comp_xcor("total")
)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metric-new}{\code{Metric$new()}}
\item \href{#method-Metric-build}{\code{Metric$build()}}
\item \href{#method-Metric-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-new}{}}}
\subsection{Method \code{new()}}{
Create a new Metric object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$new(name = NULL, func = NULL, params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String. Name of the metric.}

\item{\code{func}}{Function. Metric specification function.}

\item{\code{params}}{List. Parameters for the metric function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-build"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-build}{}}}
\subsection{Method \code{build()}}{
Build the metric specification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$build()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Metric specification object for use in simulations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
