% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_matching}
\alias{optim_matching}
\title{Determine the optimal matching for a given generation.}
\usage{
optim_matching(
  sol,
  psi_vec,
  cf_idx,
  n_iter = 10000L,
  alpha_temp = 1,
  temp0 = 1,
  eval = FALSE
)
}
\arguments{
\item{sol}{Initial solution to start annealing algorithm from}

\item{psi_vec}{Phenotype pairs for avoidance function}

\item{cf_idx}{Columns to be exchanged when generating proposal states}

\item{n_iter}{Number of iterations of annealing algorithm}

\item{alpha_temp}{Temperature decay parameter for annealing algorithm}

\item{eval}{Return simulation data for verification}
}
\value{
Final state generated by simulated annealing algorithm
}
\description{
Determine the optimal matching for a given generation.
}
