% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_comp_mean}
\alias{pheno_comp_mean}
\title{Phenotype component means metric}
\usage{
pheno_comp_mean(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Phenotype component means metric
}
