% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_comp_var}
\alias{pheno_comp_var}
\title{Phenotype component variances metric}
\usage{
pheno_comp_var(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Phenotype component variances metric
}
