% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_latent_comp_cor}
\alias{pheno_latent_comp_cor}
\title{Latent phenotype component correlation metric}
\usage{
pheno_latent_comp_cor(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Latent phenotype component correlation metric
}
