% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_latent_comp_mean}
\alias{pheno_latent_comp_mean}
\title{Latent phenotype component means metric}
\usage{
pheno_latent_comp_mean(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Latent phenotype component means metric
}
