% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_latent_comp_var}
\alias{pheno_latent_comp_var}
\title{Latent phenotype component variances metric}
\usage{
pheno_latent_comp_var(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Latent phenotype component variances metric
}
