% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{Phenotype}
\alias{Phenotype}
\title{Phenotype Class for Genetic Simulation}
\description{
The \code{Phenotype} R6 class defines and manages phenotypic traits in genetic
simulations. It allows specification of trait heritability and causal
genetic architecture through either a defined number of random causal loci
or explicit causal SNPs with their associated effect sizes.
}
\details{
The Phenotype class supports two modes of operation: (1) specifying the
number of causal loci for random selection, or (2) providing explicit
causal SNPs with their effect sizes. The class computes phenotypic values
by scaling SNP data and calculating the genetic component based on the
specified causal architecture. The heritability parameter controls the
proportion of phenotypic variance explained by genetic factors.
}
\section{Methods}{

\itemize{
\item \code{initialize()}: Create a new phenotype with specified parameters.
\item \code{compute_values()}: Calculate phenotypic values from SNP matrix data.
}
}

\seealso{
Other related classes and functions: \code{Population}, \code{MatingModel}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Get the phenotype name.}

\item{\code{values}}{Get the values for the phenotype.}

\item{\code{n_loci}}{Get the number of causal loci.}

\item{\code{causal_snps}}{Get the indices of causal SNPs.}

\item{\code{causal_effects}}{Get the causal effects.}

\item{\code{heritability}}{Get the narrow-sense heritability.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Phenotype-new}{\code{Phenotype$new()}}
\item \href{#method-Phenotype-compute_values}{\code{Phenotype$compute_values()}}
\item \href{#method-Phenotype-clone}{\code{Phenotype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-new"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new Phenotype object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$new(
  name,
  heritability,
  n_loci = NULL,
  causal_snps = NULL,
  causal_effects = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String. Name of the phenotype.}

\item{\code{heritability}}{Numeric. The narrow-sense heritability of the
phenotype.}

\item{\code{n_loci}}{Integer. Number of causal loci affecting the phenotype.}

\item{\code{causal_snps}}{Integer vector of length equal to \code{n_loci} specifying
the indices of causal SNPs.}

\item{\code{causal_effects}}{Numeric vector of length equal to \code{n_loci}
specifying the effect sizes of causal SNPs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-compute_values"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-compute_values}{}}}
\subsection{Method \code{compute_values()}}{
Compute the data vector of a Phenotype.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$compute_values(snp_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{snp_matrix}}{Matrix of SNPs from a Population instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
