% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Population}
\alias{Population}
\title{Population class}
\description{
Population class

Population class
}
\details{
Represents a population for performing simulations of xAM, allowing for
configuration via constructor arguments.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Data frame containing SNP and phenotype data}

\item{\code{n_pop}}{Number of individuals in the population}

\item{\code{n_loci}}{Number of genetic loci (SNPs) modelled in the genome}

\item{\code{phenotypes}}{A list of Phenotype objects}

\item{\code{maf_dist}}{One of 'beta' or 'uniform'}

\item{\code{maf_params}}{Parameters for MAF distribution}

\item{\code{maf}}{Minor allele frequencies for SNPs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Population-new}{\code{Population$new()}}
\item \href{#method-Population-add_phenotypes}{\code{Population$add_phenotypes()}}
\item \href{#method-Population-remove_phenotypes}{\code{Population$remove_phenotypes()}}
\item \href{#method-Population-clone}{\code{Population$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-new"></a>}}
\if{latex}{\out{\hypertarget{method-Population-new}{}}}
\subsection{Method \code{new()}}{
Instantiate a new \code{Population} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$new(
  n_pop,
  n_loci,
  maf_dist = NULL,
  maf_params = NULL,
  phenotypes = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_pop}}{Number of individuals in the population.}

\item{\code{n_loci}}{Number of loci in the genome.}

\item{\code{maf_dist}}{Minor allele frequency distribution type ("uniform" or
"beta").}

\item{\code{maf_params}}{Parameters for the minor allele frequency distribution.}

\item{\code{phenotypes}}{List of Phenotype objects to model in the population.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Population} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-add_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-Population-add_phenotypes}{}}}
\subsection{Method \code{add_phenotypes()}}{
Add new phenotypes to model in the population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$add_phenotypes(phenotypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotypes}}{A list of Phenotype objects to add to the population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-remove_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-Population-remove_phenotypes}{}}}
\subsection{Method \code{remove_phenotypes()}}{
Remove phenotypes modelled in the population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$remove_phenotypes(phenotypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotypes}}{Names of phenotypes to remove from the population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Population-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
