% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{process_config}
\alias{process_config}
\title{Process the configuration file}
\usage{
process_config(config)
}
\arguments{
\item{config}{The configuration object to process}
}
\value{
A list object with data processed from the configuration file
}
\description{
Extracts and transforms data registered in configuration file
to be used in the simulation.
}
