% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi}
\alias{psi}
\title{Evaluate the energy of a state}
\usage{
psi(rho_sol, rho)
}
\arguments{
\item{sol}{The current state matrix}

\item{swap}{Indices of swapped pairs}

\item{phi_m}{Male phenotype indices}

\item{phi_f}{Female phenotype indices}
}
\value{
Vector of Δ values for each phenotype pair
}
\description{
Evaluate the energy of a state
}
