% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_pop}
\alias{simulate_pop}
\title{Simulate a population from a config file}
\usage{
simulate_pop(config_path, progress = FALSE)
}
\arguments{
\item{config_path}{Path to configuration file with simulation parameters}

\item{quietly}{Display progress bar when running simulation}
}
\value{
A list which population, snp_pair_cors, and sib_cors objects.
}
\description{
Simulate a population from a config file
}
