% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_population}
\alias{simulate_population}
\title{Simulate a population from a config file}
\usage{
simulate_population(config_path, init_population = NULL)
}
\arguments{
\item{config_path}{Path to configuration file with simulation parameters}
}
\value{
An amsim object containing the results of the simulation
}
\description{
Simulate a population from a config file
}
