% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Simulation}
\alias{Simulation}
\title{Simulation class}
\description{
Class in the amsimr package to perform assortative mating simulations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{population}}{Population object representing the initial population.}

\item{\code{summariser}}{Summariser object to compute summary data each generation.}

\item{\code{mating_model}}{MatingModel object to model the population mating regime.}

\item{\code{n_generations}}{Number of generations to simulate.}

\item{\code{n_iterations}}{Number of simulation iterations to perform.
Instantiate a new Simulation object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulation-new}{\code{Simulation$new()}}
\item \href{#method-Simulation-run}{\code{Simulation$run()}}
\item \href{#method-Simulation-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(
  population,
  summariser,
  mating_model,
  n_generations,
  n_iterations = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{Population object representing the initial population.}

\item{\code{summariser}}{Summariser object to compute summary data each generation.}

\item{\code{mating_model}}{MatingModel object to model the population mating regime.}

\item{\code{n_generations}}{Number of generations to simulate.}

\item{\code{n_iterations}}{Number of simulation iterations to perform.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-run"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
