% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{transform_summary_data}
\alias{transform_summary_data}
\title{Transform data compiled by summarisers for an amsim object.}
\usage{
transform_summary_data(summary_data)
}
\arguments{
\item{summary_data}{The amsim object}
}
\value{
Transformed summary data
}
\description{
Transform data compiled by summarisers for an amsim object.
}
