% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validate the entire configuration file}
\usage{
validate_config(config)
}
\arguments{
\item{config}{The configuration object to validate}
}
\description{
Validates that the configuration file has all required sections and
that all required parameters have valid values.
}
