% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_section}
\alias{validate_section}
\title{Validate a section of the configuration against rules}
\usage{
validate_section(config, section, rules)
}
\arguments{
\item{config}{The configuration object}

\item{section}{The section name to validate}

\item{rules}{A list of validation rules}
}
\description{
Validates a section of the simulation configuration file, i.e. whether
a key within the section is required and how to validate it, using a
validation function.
}
