% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_delta}
\alias{compute_delta}
\title{Computes correlation update when swapping pairs}
\usage{
compute_delta(sol_mat, swap_idx, male_snp_idx, female_snp_idx)
}
\arguments{
\item{sol_mat}{Solution matrix}

\item{swap_idx}{Indices to swap}

\item{male_snp_idx}{Male SNP indices}

\item{female_snp_idx}{Female SNP indices}
}
\value{
Correlation vector update
}
\description{
Computes correlation update when swapping pairs
}
