% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_dpsi}
\alias{compute_dpsi}
\title{Computes energy differential between states}
\arguments{
\item{curr_cor}{Current correlation}

\item{target_cor}{Target correlation}

\item{delta_cor}{Correlation update}
}
\value{
Energy differential
}
\description{
Computes energy differential between states
}
\keyword{internal}
