% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_psi}
\alias{compute_psi}
\title{Evaluates energy of current state}
\arguments{
\item{curr_cor}{Current correlation}

\item{target_cor}{Target correlation}
}
\value{
Energy value
}
\description{
Evaluates energy of current state
}
\keyword{internal}
