% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{custom_metric}
\alias{custom_metric}
\title{Create a custom Metric}
\usage{
custom_metric(
  name,
  metric_func,
  n_rows = NULL,
  n_cols = NULL,
  labels = NULL,
  require_lat = FALSE
)
}
\arguments{
\item{name}{Name of the custom metric function}

\item{metric_func}{The R function defining the metric}

\item{n_rows}{Number of rows in output from single \code{metric_func} call}

\item{n_cols}{Number of columns in output from single \code{metric_func} call}

\item{labels}{Labels of output data produced by single \code{metric_func} call}

\item{require_lat}{Does this metric require a latent phenotype buffer?}
}
\description{
Create a custom Metric
}
