% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_matching}
\alias{optim_matching}
\title{Performs simulated annealing to optimize matching}
\usage{
optim_matching(
  sol_mat,
  snp_pairs,
  female_swap_idx,
  num_iterations = 10000L,
  temp_decay = 1,
  init_temp = 1,
  collect_metrics = FALSE
)
}
\arguments{
\item{sol_mat}{Solution matrix to optimize}

\item{snp_pairs}{SNP pair data \link{male_idx, female_idx, target_cor}}

\item{female_swap_idx}{Columns to swap in solution}

\item{temp_decay}{Temperature decay rate}

\item{init_temp}{Initial temperature}

\item{collect_metrics}{Whether to collect optimization metrics}

\item{max_iterations}{Maximum iterations to run}
}
\value{
Optimized solution matrix and optional metrics
}
\description{
Performs simulated annealing to optimize matching
}
