% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_matching}
\alias{optim_matching}
\title{Performs simulated annealing to optimize matching}
\usage{
optim_matching(
  sol_mat,
  snp_pairs,
  female_swap_idx,
  num_iterations = 10000L,
  temp_decay = 0.995,
  init_temp = 1e-09,
  auto_temp_samples = 100000L,
  auto_accept_ratio = 0.995,
  collect_diagnostics = FALSE,
  quietly = TRUE
)
}
\arguments{
\item{sol_mat}{Solution matrix to optimize}

\item{snp_pairs}{SNP pair data \link{male_idx, female_idx, target_cor}}

\item{female_swap_idx}{Columns to swap in solution}

\item{temp_decay}{Temperature decay rate}

\item{init_temp}{Initial temperature}

\item{auto_temp_samples}{Number of samples to draw for determining initial
temperature of simulated annealing algorithm}

\item{auto_accept_ratio}{Desired initial acceptance ratio for calibration of
initial temperature value in simulated annealing algorithm}

\item{collect_diagnostics}{If TRUE, collect diagnostic metrics}

\item{quietly}{If TRUE, suppress output messages}

\item{max_iterations}{Maximum iterations to run}
}
\value{
Optimized solution matrix and optional metrics
}
\description{
Performs simulated annealing to optimize matching
}
