% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_comp_cor}
\alias{pheno_comp_cor}
\title{Phenotype component correlations metric}
\usage{
pheno_comp_cor(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Phenotype component correlations metric
}
