% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_comp_xcor}
\alias{pheno_comp_xcor}
\title{Between-mate phenotype component cross-correlation metric}
\usage{
pheno_comp_xcor(component_type)
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Between-mate phenotype component cross-correlation metric
}
