% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pheno_latent_h2}
\alias{pheno_latent_h2}
\title{Latent phenotype heritabilities metric}
\usage{
pheno_latent_h2()
}
\arguments{
\item{component_type}{Target component type.}
}
\description{
Latent phenotype heritabilities metric
}
