% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{Phenotype}
\alias{Phenotype}
\title{Phenotype Class}
\description{
Represents a phenotype with its associated genetic properties.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of phenotype}

\item{\code{loci}}{Indices of causal loci}

\item{\code{n_loci}}{Number of causal loci}

\item{\code{loci_effects}}{Genic effects of causal loci}

\item{\code{heritability}}{Narrow-sense heritability of phenotype
Initialise a Phenotype object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Phenotype-new}{\code{Phenotype$new()}}
\item \href{#method-Phenotype-set_loci}{\code{Phenotype$set_loci()}}
\item \href{#method-Phenotype-clone}{\code{Phenotype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-new"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$new(
  name,
  heritability,
  loci = NULL,
  n_loci = NULL,
  loci_effects = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the phenotype}

\item{\code{heritability}}{Narrow-sense heritability of phenotype}

\item{\code{n_loci}}{Number of causal loci}

\item{\code{loci_effects}}{Vector of genic effects of causal loci}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-set_loci"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-set_loci}{}}}
\subsection{Method \code{set_loci()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$set_loci(population_n_loci, indices = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Phenotype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Phenotype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Phenotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
