% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_populations}
\alias{simulate_populations}
\title{Simulate a number of population from a config file, with parallel options}
\usage{
simulate_populations(config_path, n_sims = 10, progress = FALSE)
}
\arguments{
\item{config_path}{Path to configuration file with simulation parameters}

\item{n_sims}{Number of populations to simulate before aggregating data}

\item{progress}{Display progress bar when running simulation}
}
\value{
A list which population, snp_pair_cors, and sib_cors objects.
}
\description{
Simulate a number of population from a config file, with parallel options
}
